# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mini RBAC Go API",
    "version": "2.0.0",
    "description": "Role-Based Access Control API for managing roles, groups, permissions, and workspace hierarchies",
    "contact": {
      "name": "Mini RBAC Go",
      "url": "https://github.com/redhat/mini-rbac-go"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8085/api/rbac/v2",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/workspaces": {
      "get": {
        "tags": ["Workspaces"],
        "summary": "List all workspaces",
        "description": "Returns a paginated list of workspaces for the tenant",
        "operationId": "listWorkspaces",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Workspaces"],
        "summary": "Create a new workspace",
        "description": "Creates a new workspace with optional parent",
        "operationId": "createWorkspace",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workspace created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/workspaces/{id}": {
      "get": {
        "tags": ["Workspaces"],
        "summary": "Get workspace by ID",
        "description": "Returns a single workspace, optionally with ancestry information",
        "operationId": "getWorkspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Workspace UUID"
          },
          {
            "name": "include_ancestry",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include ancestry field with all ancestor workspaces"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": ["Workspaces"],
        "summary": "Update workspace",
        "description": "Updates workspace name and description",
        "operationId": "updateWorkspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Workspace UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Workspaces"],
        "summary": "Delete workspace",
        "description": "Deletes a workspace (must not have children)",
        "operationId": "deleteWorkspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Workspace UUID"
          }
        ],
        "responses": {
          "204": {
            "description": "Workspace deleted successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/workspaces/{id}/move": {
      "post": {
        "tags": ["Workspaces"],
        "summary": "Move workspace to a new parent",
        "description": "Moves a workspace under a new parent. Validates hierarchy constraints.",
        "operationId": "moveWorkspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Workspace UUID to move"
          },
          {
            "name": "parent_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "UUID of the new parent workspace"
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace moved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "description": "Number of items in current response"
          }
        }
      },
      "Workspace": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Workspace UUID"
          },
          "name": {
            "type": "string",
            "description": "Workspace name",
            "example": "Engineering"
          },
          "description": {
            "type": "string",
            "description": "Workspace description",
            "example": "Engineering department workspace"
          },
          "type": {
            "type": "string",
            "enum": ["root", "default", "standard"],
            "description": "Workspace type"
          },
          "parent_id": {
            "type": "string",
            "format": "uuid",
            "description": "Parent workspace UUID",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "modified": {
            "type": "string",
            "format": "date-time",
            "description": "Last modification timestamp"
          }
        }
      },
      "CreateWorkspaceRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Workspace name",
            "example": "Engineering"
          },
          "description": {
            "type": "string",
            "description": "Workspace description",
            "example": "Engineering department workspace"
          },
          "type": {
            "type": "string",
            "enum": ["root", "default", "standard"],
            "description": "Workspace type (default: standard)",
            "default": "standard"
          },
          "parent_id": {
            "type": "string",
            "format": "uuid",
            "description": "Parent workspace UUID (optional)"
          }
        }
      },
      "UpdateWorkspaceRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Workspace name",
            "example": "Updated Engineering"
          },
          "description": {
            "type": "string",
            "description": "Workspace description",
            "example": "Updated description"
          }
        }
      },
      "WorkspaceListResponse": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workspace"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Error title"
          },
          "detail": {
            "type": "string",
            "description": "Detailed error message"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          }
        }
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}